--- app/models/delivery_receipt.rb
@@ -1,4 +1,0 @@
-class DeliveryReceipt < ActiveRecord::Base
-  auto_increment :number
-  has_one :contract
-end
