+++ config/initializers/date_format.rb
@@ -1,2 +1,5 @@
 Time::DATE_FORMATS[:default] = "%d/%m/%Y"
 Date::DATE_FORMATS[:default] = "%d/%m/%Y"
+
+Time::DATE_FORMATS[:written_date] = "%d de %B de %Y"
+Date::DATE_FORMATS[:written_date] = "%d de %B de %Y"
