--- db/migrate/20151028233129_create_delivery_receipts.rb
@@ -1,9 +1,0 @@
-class CreateDeliveryReceipts < ActiveRecord::Migration
-  def change
-    create_table :delivery_receipts do |t|
-      t.integer :number
-
-      t.timestamps null: false
-    end
-  end
-end
