--- spec/factories/delivery_receipts.rb
@@ -1,6 +1,0 @@
-FactoryGirl.define do
-  factory :delivery_receipt do
-    number 1
-  end
-
-end
